SmartResumeAnalyzer/requirements.txt
streamlit
python-dotenv
pdf2image
Pillow
google-generativeai
